Change tests from nose to pytest
Some of the 'yield' tests are easy to refactor into parameterised tests
while others would take a fair amount of rethinking and/or additional
helper functions. At this stage, the relevant functions are simply decorated
with a 'collect_yielded' decorator that collects all the yield tests at
collection time and sets them up as a parameterised test to be run later.