Move blackbox tests to pytest
The approach is to add a file collection hook that generates test objects based
on every .po (etc) file that it sees in the directory. The test objects then
know how to actually run the test on the .po file, ensuring that the appropriate
tags are emitted.

The two coded tests (test_os_error_no_such_file, test_os_error_permission_denied)
need to be moved out of the pytest configuration (conftest.py) so that pytest
finds them and adds them to the list of tests to run.